bot-estiv/apps/api/pyproject.toml
[project]
name = "bot-estiv-api"
version = "0.1.0"
description = "Bot Estiv - Multiagente de Marketing Digital para Gardens Wood"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.2.0",
    "alembic>=1.13.2",
    "pgvector>=0.3.0",
    "redis>=5.0.7",
    "arq>=0.26.1",
    "langgraph>=0.2.28",
    "langgraph-checkpoint-postgres>=1.0.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "langchain-community>=0.3.0",
    "google-genai>=0.3.0",
    "twilio>=9.2.3",
    "facebook-business>=19.0.0",
    "httpx>=0.27.0",
    "boto3>=1.35.0",
    "pillow>=10.4.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.14.0",
    "ffmpeg-python>=0.2.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bot_estiv"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
